+++ app/jobs/disco_app/concerns/app_installed_job.rb
@@ -14,10 +14,10 @@ module DiscoApp::Concerns::AppInstalledJob
   # - Perform initial update of shop information.
   # - Subscribe to default plan, if any exists.
   #
-  def perform(shopify_domain, plan_code = nil, source = nil)
-    DiscoApp::SynchroniseWebhooksJob.perform_now(shopify_domain)
-    DiscoApp::SynchroniseCarrierServiceJob.perform_now(shopify_domain)
-    DiscoApp::ShopUpdateJob.perform_now(shopify_domain)
+  def perform(shop, plan_code = nil, source = nil)
+    DiscoApp::SynchroniseWebhooksJob.perform_now(@shop)
+    DiscoApp::SynchroniseCarrierServiceJob.perform_now(@shop)
+    DiscoApp::ShopUpdateJob.perform_now(@shop)
 
     @shop.reload